refactor: rename package name into autoware_mtr
