+++ include/autoware/mtr/agent.hpp
@@ -23,7 +23,7 @@
 #include <tuple>
 #include <vector>
 
-namespace autoware::trt_mtr
+namespace autoware::mtr
 {
 constexpr size_t AgentStateDim = 12;
