+++ include/autoware/mtr/polyline.hpp
@@ -24,7 +24,7 @@
 #include <tuple>
 #include <vector>
 
-namespace autoware::trt_mtr
+namespace autoware::mtr
 {
 constexpr size_t PointStateDim = 7;
