+++ include/autoware/mtr/trajectory.hpp
@@ -20,7 +20,7 @@
 #include <cstddef>
 #include <vector>
 
-namespace autoware::trt_mtr
+namespace autoware::mtr
 {
 
 constexpr size_t PredictedStateDim = 7;
