+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>autoware_tensorrt_mtr</name>
+  <name>autoware_mtr</name>
   <version>0.1.0</version>
   <description>ROS 2 Node of Motion Transfromer(a.k.a MTR).</description>
   <maintainer email="kotaro.uetake@tier4.jp">kotarouetake</maintainer>
