+++ src/builder.cpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/tensorrt_mtr/builder.hpp"
+#include "autoware/mtr/builder.hpp"
 
 #include <fstream>
 #include <sstream>
 
-namespace autoware::trt_mtr
+namespace autoware::mtr
 {
 namespace
 {
