+++ src/node.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/tensorrt_mtr/node.hpp"
+#include "autoware/mtr/node.hpp"
 
 #include <lanelet2_extension/utility/message_conversion.hpp>
 
@@ -27,7 +27,7 @@
 #include <algorithm>
 #include <cmath>
 
-namespace autoware::trt_mtr
+namespace autoware::mtr
 {
 namespace
 {
@@ -157,7 +157,7 @@ CalibrationType getCalibrationType(const std::string & name)
 }  // namespace
 
 MTRNode::MTRNode(const rclcpp::NodeOptions & node_options)
-: rclcpp::Node("tensorrt_mtr", node_options), transform_listener_(this), polyline_type_map_(this)
+: rclcpp::Node("mtr", node_options), transform_listener_(this), polyline_type_map_(this)
 {
   // TODO(ktro2828)
   {
@@ -649,4 +649,4 @@ PredictedObject MTRNode::generatePredictedObject(
 }  // namespace autoware::trt_mtr
 
 #include <rclcpp_components/register_node_macro.hpp>
-RCLCPP_COMPONENTS_REGISTER_NODE(autoware::trt_mtr::MTRNode);
+RCLCPP_COMPONENTS_REGISTER_NODE(autoware::mtr::MTRNode);
